requirements.txt
# Core ML/AI
transformers==4.36.0
torch==2.1.0
pandas==2.1.4
numpy==1.24.3

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Data Processing
beautifulsoup4==4.12.2
PyPDF2==3.0.1
openpyxl==3.1.2
python-dotenv==1.0.0

# Email/Cloud
imaplib2==3.6
email-validator==2.1.0
boto3==1.34.0  # For AWS

# Database
sqlalchemy==2.0.23
sqlite3  # Built-in

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Alerts
secure-smtplib==0.1.1

# API (optional)
fastapi==0.104.1
uvicorn==0.24.0